Moved token calls out of workspace repo
